--- a/.travis.yml
+++ b/.travis.yml
@@ -1,0 +1,31 @@
+language: python
+
+python:
+ - pypy
+ - pypy3
+ - 2.6
+ - 2.7
+ - 3.2
+ - 3.3
+ - 3.4
+
+sudo: false
+
+install:
+ - pip install coverage tweepy
+
+script:
+ - coverage run --source=everywordbot test/test_everywordbot.py -v
+
+after_success:
+ - pip install coveralls
+ - coveralls
+
+after_script:
+ - coverage report
+ - pip install pep8 pyflakes
+ - pep8 --statistics --count *.py test/*.py
+ - pyflakes *.py test/*.py | tee >(wc -l)
+
+matrix:
+  fast_finish: true