--- a/.travis.yml
+++ b/.travis.yml
@@ -23,8 +23,8 @@ after_success:
 after_script:
  - coverage report
  - pip install pep8 pyflakes
- - pep8 --statistics --count *.py test/*.py
- - pyflakes *.py test/*.py | tee >(wc -l)
+ - pep8 --statistics --count .
+ - pyflakes . | tee >(wc -l)
 
 matrix:
   fast_finish: true