--- a/test/test_source.txt
+++ b/test/test_source.txt
@@ -1,0 +1,3 @@
+word1
+word2
+word3