Includes PR #2.

I tested five different implementations like this:

 * Modified test_source.py to add 94,110 words (one per line). File is 1.2 MB.
 * Modified `test__get_current_line()` in test_everywordbot.py to call `line = bot._get_current_line(index)` 100 times and assert the word at index 93,363.
 * Run the tests `time python test/test_everywordbot.py` and check the real time (Python 2.7/Windows 7). 
 * Run two more times, average.

I didn't commit these test changes, but the different implementations were committed for comparison.

The original implementation took 4.2 seconds. Other implementations reduced this to 3.0s, 1.3s, 1.2s and 0.2s.

The fastest by far is to use [`linecache`](https://docs.python.org/2/library/linecache.html):
```python
return linecache.getline(self.source_file_name, index+1).strip()
```

This reads the whole file in. One thing to note is:

> This function will never raise an exception — it will return `''` on errors 

Is this a problem? I expect that would instead cause Tweepy to raise an exception instead of _get_current_line(), so just shifting it a bit.