@aparrish 2015-05-21
I might be mistaken, but I think the performance improvement with `linecache` only applies if you're accessing the file *more than once* during the execution of the program, right? That isn't the case with this program—it only accesses the file once per execution. A better timing test would be to create an `EverywordBot` object and call its `_get_current_line` method, destroy it, and repeat that 100 times, to better approximate this use case.

(I also don't like that `linecache` does all kinds of weird magic in searching for the file—I'm worried that this might lead to unexpected behavior for some users.)

I like the proposals in the other commits—I'm surprised actually that just getting rid of `.strip()` gives a 25% performance improvement. But I also want to keep memory usage down, since one of the ways I'm hoping people will be able to use this script is with very large text files on a VPS with very little memory. Could you resubmit a pull request that takes this requirement into account? (The `.readlines()` solution obviously uses a lot of memory, since it reads in the entire file, but I'm not sure off the top of my head if the `enumerate` solution will do the same.)