@hugovk 2015-05-21
Ah yes, you're right the test wasn't fair due to re-reading the same file 100 times. I'll check it and make a new PR. I agree about avoiding linecache's weird magic.